% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econocharts-package.R
\docType{package}
\name{econocharts-package}
\alias{econocharts-package}
\title{econocharts: Microeconomic Made with 'ggplot2'}
\description{
This package creates...
}
\details{
\itemize{
\item{Package: econocharts}
\item{Version: 1.0}
\item{Maintainer: José Carlos Soage González \email{jsoage@uvigo.es}}
}
}
\seealso{
\itemize{
\item{\href{https://r-coder.com/}{R tutorials}}
}
}
\author{
\itemize{
\item{Soage González, José Carlos.}
\item{Weiss, Andrew.}
}
}
