% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppf.R
\name{ppf}
\alias{ppf}
\title{Production–possibility frontier}
\usage{
ppf(
  ...,
  xmax,
  ymax,
  type = "concave",
  x,
  linecol,
  labels,
  generic = TRUE,
  geom = "text",
  geomcol = 1,
  geomfill = "white",
  main = NULL,
  sub = NULL,
  xlab = NULL,
  ylab = NULL,
  acol,
  alpha = 0.3,
  bg.col = "white"
)
}
\arguments{
\item{...}{Specify the production–possibility frontiers separated by comma (as \code{data.frame}) you want to display in the graph. This will override the sample curve.}

\item{xmax}{Numeric. Allows modifying the maximum X value for the default production–possibility frontier.}

\item{ymax}{Numeric. Allows modifying the maximum Y value for the default production–possibility frontier.}

\item{type}{Possible values are \code{"concave"} (default) and \code{"line"} to plot a concave or a linear production–possibility frontier function by default, respectively.}

\item{x}{Y-axis values where to create intersections with the production–possibility frontier}

\item{linecol}{Line color of the curves.}

\item{labels}{If \code{x} is specified are the labels for the intersection points.}

\item{generic}{Boolean. If \code{TRUE} and \code{x} is specified, the axis labels shows generic names. If \code{FALSE}, the axis labels are the actual data of the axis that corresponds to the intersection points.}

\item{geom}{Possible values are \code{"text"} to display the labels of the intersection points with text and \code{"label"} to show them with labels.}

\item{geomcol}{Color of the labels of the intersection points.}

\item{geomfill}{If \code{geom = "label"} is the background color of the label.}

\item{main}{Main title of the plot.}

\item{sub}{Subtitle of the plot.}

\item{xlab}{Name of the X-axis.}

\item{ylab}{Name of the Y-axis.}

\item{acol}{Color of the area of the below the production–possibility frontier}

\item{alpha}{Transparency of the colored area}

\item{bg.col}{Background color of the plot}
}
\description{
Creates production–possibility frontiers. The function allows specifying custom frontiers, modifying the type of the curves (concave or linear), creating intersections along the values of the Y-axis and the curve and customizing the final output with further arguments.
}
