% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptvalue.R
\name{ptvalue}
\alias{ptvalue}
\title{Value function in Prospect Theory}
\usage{
ptvalue(
  x,
  sigma = 0.3,
  lambda = -2.25,
  xint,
  xintcol = 1,
  main = NULL,
  sub = NULL,
  xlab = "Loss / Gain",
  ylab = "Value",
  col = 1,
  bg.col = "white",
  ticks = TRUE,
  xlabels = TRUE,
  ylabels = TRUE,
  by_x = 10,
  by_y = 20
)
}
\arguments{
\item{x}{Numeric. Vector of gain / loss values for x.}

\item{sigma}{Number. Exponent of functions, should be less than 1 to make an
'S' shaped curve.}

\item{lambda}{Number. Extent of asymmetry between losses and gains. Should be
less than -1 for losses to be more 'intense' than gains (as suggested by
Prospect Theory). Between -1 and 0 for gains to be more intense than
losses. Greater than 0 for losses to have positive value.}

\item{xint}{Numeric. Symmetric intersections. X-intercept values where to
highlight points -- will be placed at both \code{xint} and \code{abs(xint)} to
demonstrate asymmetry in \code{value}.}

\item{xintcol}{Color of dashed lines calling-out \code{xint}.}

\item{main}{Main title of the plot.}

\item{sub}{Subtitle of the plot.}

\item{xlab}{Name of the X-axis.}

\item{ylab}{Name of the Y-axis.}

\item{col}{Color of function segment.}

\item{bg.col}{Background color.}

\item{ticks}{TOOD}

\item{xlabels}{TRUE / FALSE :  whether x labels are included.}

\item{ylabels}{TRUE / FALSE :  whether y labels are included.}

\item{by_x}{Number. Increment of the x-axis labels.}

\item{by_y}{Number. Increment of the x-axis labels.}
}
\description{
Produces asymmetric S-shaped value function according to lessons
from Prospect Theory that losses are felt more intensely than gains.
}
\details{
TODO
}
\examples{

ptvalue(
  sigma = 0.25,
  xint = 20,
  xintcol = 'blue',
  main = "Prospect Theory Shows That Gains & Losses are Felt Assymetrically",
  sub = "Losses are More Intense"
)

}
\references{
Tversky, Amos; Kahneman, Daniel (1992). "Advances in prospect
theory: Cumulative representation of uncertainty". Journal of Risk and
Uncertainty. 5 (4): 297–323.
}
