% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdcurve.R
\name{sdcurve}
\alias{sdcurve}
\title{Supply and demand chart}
\usage{
sdcurve(
  ...,
  xmax,
  ymax,
  generic = TRUE,
  equilibrium = TRUE,
  main = NULL,
  sub = NULL,
  xlab = NULL,
  ylab = NULL,
  curve_names = TRUE,
  labels,
  col,
  bg.col = "white"
)
}
\arguments{
\item{bg.col}{}
}
\description{
econocharts
}
