% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax.R
\name{tax_graph}
\alias{tax_graph}
\title{Tax graph}
\usage{
tax_graph(
  demand_fun,
  supply_fun,
  supply_tax,
  names = c("Consumer surplus", "Producer surplus", "DWL", "Consumer tax burden",
    "Producer tax burden"),
  title = NULL,
  xlab = "Product (Q)",
  ylab = "Price (P)",
  colors,
  shaded = FALSE,
  xlim = c(0, 45),
  ylim = c(0, 20),
  max_x = 45
)
}
\arguments{
\item{demand_fun}{TODO}

\item{supply_fun}{TODO}

\item{supply_tax}{TODO}

\item{names}{TODO}

\item{title}{TODO}

\item{xlab}{TODO}

\item{ylab}{TODO}

\item{colors}{TODO}

\item{shaded}{TODO}

\item{xlim}{TODO}

\item{ylim}{TODO}

\item{max_x}{TODO}
}
\description{
TODO
}
\examples{

# Data
demand <- function(Q) 20 - 0.5 * Q
demand_new <- function(Q) demand(Q) + 5
supply <- function(Q) 2 + 0.25 * Q
supply_new <- function(Q) supply(Q) + 5

supply_tax <- function(Q) supply(Q) + 5

demand_elastic <- function(Q) 10 - 0.05 * Q
demand_inelastic <- function(Q) 20 - 2 * Q

supply_elastic <- function(Q) 2 + 0.05 * Q
supply_elastic_tax <- function(Q) supply_elastic(Q) + 5
supply_inelastic <- function(Q) 2 + 1.5 * Q
supply_inelastic_tax <- function(Q) supply_inelastic(Q) + 5

normal_taxes <- tax_graph(demand, supply, supply_tax, NULL)
normal_taxes_shaded <- tax_graph(demand, supply, supply_tax, NULL, shaded = TRUE)
normal_taxes_shaded$p

}
\author{
\itemize{
\item{Weiss, Andrew.}
}
}
