% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptvalue.R
\name{ptvalue}
\alias{ptvalue}
\title{Value function in Prospect Theory}
\usage{
ptvalue(
  x,
  sigma = 0.88,
  lambda = -2.25,
  xint,
  xintcol = 1,
  main = NULL,
  sub = NULL,
  xlab = "Loss",
  ylab = "Value",
  col = 1,
  bg.col = "white",
  ticks = TRUE,
  xlabels = TRUE,
  ylabels = TRUE,
  by_x = 10,
  by_y = 20
)
}
\arguments{
\item{x}{TODO}

\item{sigma}{TODO}

\item{lambda}{TODO}

\item{xint}{TODO}

\item{xintcol}{TODO}

\item{main}{TODO}

\item{sub}{TODO}

\item{xlab}{TODO}

\item{ylab}{TODO}

\item{col}{TODO}

\item{bg.col}{TODO}

\item{ticks}{TOOD}

\item{xlabels}{TODO}

\item{ylabels}{TODO}

\item{by_x}{TODO}

\item{by_y}{TODO}
}
\description{
TODO
}
\details{
TODO
}
\examples{

# Sigma
sigma <- 0.25

# Symmetric intersections
xint <- seq(10, 80, 10)
ptvalue(sigma = sigma, col = 1, xint = xint, xintcol = 4)

}
